Fix a crash in libgcrypt caused by the new linker in clang 5
When libgcrypt is built using clang 5 it crashes when running tests and later in runtime when for example signing a message with a 4096 rsa key.
To solve this issue it is necessary to disable the new linker introduced in clang 5 by passing the `-ld_classic` flag to the linker arguments.